tools: copilot: rewrite copilot to prepare future optimizations
Remove singleton to be able to use multiples workers with gunicorn+uvicorn
(not possible yet).
Use application factory pattern (create_app and app in app.py) to support
running copilot with gunicorn+uvicorn.
Remove command line options usage not compatible with gunicorn+uvicorn.
Add Settings class to set default parameters and modify them using
environment variables.
