+++ docs/usage/copilot.md
@@ -8,16 +8,26 @@ to this server and receive commands from monitors.
 It provides a web server to provide monitoring on any devices (PC, smartphones)
 connected to the same network.
 
-## Command line options
+## Run Copilot
 
 ```bash
-$ cogip-copilot --help
-Usage: cogip-copilot [OPTIONS]
+$ cogip-copilot
+```
 
-Launch COGIP Copilot.
+## Parameters
 
-Options:
-  -s, --serial PATH               Serial port connected to STM32 device  [default: /dev/ttyUSB0]
-  -b, --baud INTEGER              Baud rate  [default: 230400]
-  -p, --port INTEGER              Socket.IO/Web server port  [default: 80]
+Copilot default parameters can be modified using environment variables.
+All variables can be defined in the `.env` file.
+
+Example of `.env` file with all default values:
+
+```bash
+# Socket.IO/Web server port
+COPILOT_SERVER_PORT=8080
+
+# Serial port connected to STM32 device
+COPILOT_SERIAL_PORT="/dev/ttyUSB0"
+
+# Baud rate
+COPILOT_SERIAL_BAUD=230400
 ```
